Optimize system performance and enhance code readability
This commit addresses two major aspects: performance optimization and code readability enhancement. Small changes in the logic of queue document count in the 'Mover.java' and 'Processor.java' classes were made to better handle large volumes of document. Additionally, an unnecessary RefreshCaches method was removed from the 'ReadManager.java' and cache refresh logic was improved to be more consistent and efficient.

A new class 'Reprocesser.java' was introduced to handle document processing in a more streamline way. Moreover, the 'Main.java' was renamed into 'CrawlyServer.java' to better reflect its purpose. To enhance readability and maintainability, the unused imports were removed and import statements were organized.

Also, in the 'WriteManager.java' class, a switch-case structure was replaced with a direct method call to 'Database.toCollection()', making the code cleaner and easier to follow.

These changes are expected to overall enhance the performance of the service and make the codebase cleaner and more maintainable."